+++ wagtail/admin/api/v2/actions/copy_for_translation.py
