+++ wagtail/admin/api/v2/actions/delete.py
