+++ wagtail/admin/api/v2/actions/publish.py
