+++ wagtail/admin/api/v2/serializers.py
